Generator: Split into more files
